+++ minimap.c
@@ -10,6 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <math.h>
+#include <mlx.h>
+#include "cub3d.h"
+#include "cub3d_int.h"
+
 void	c3_draw_player_on_map(t_c3_state *stat)
 {
 	const char	c3_player_bitmap[] = {
@@ -86,3 +91,45 @@ void	c3_draw_map(t_c3_state *stat)
 	c3_draw_player_on_map(stat);
 }
 
+void	c3_draw_rays_on_map(t_c3_state *stat)
+{
+	double		world_x;
+	double		world_y;
+	double		screen_x;
+	double		screen_y;
+	int			x;
+	int			i;
+
+	x = 0;
+	while (x < stat->renderer.resolution_x)
+	{
+		world_x = stat->renderer.rays[x].hits[0].position.x;
+		world_y = stat->renderer.rays[x].hits[0].position.y;
+		screen_x = world_x * stat->renderer.minimap_width / stat->map.width;
+		screen_y = world_y * stat->renderer.minimap_height / stat->map.height;
+
+		int r = 255 * x / stat->renderer.resolution_x;
+		int col = (r << 16) + ((255 - r) << 0);
+		mlx_string_put(
+			stat->mlx, stat->window, screen_x, screen_y,
+			mlx_get_color_value(stat->mlx, col), "*");
+
+		i = 0;
+		while (i < stat->renderer.rays[x].hit_sprite_count)
+		{
+			world_x = stat->renderer.rays[x].hits[i + 1].position.x;
+			world_y = stat->renderer.rays[x].hits[i + 1].position.y;
+			screen_x = world_x * stat->renderer.minimap_width / stat->map.width;
+			screen_y = world_y * stat->renderer.minimap_height / stat->map.height;
+
+			int r = 255 * x / stat->renderer.resolution_x;
+			int col = (r << 16) + ((255 - r) << 0);
+			mlx_string_put(
+				stat->mlx, stat->window, screen_x, screen_y,
+				mlx_get_color_value(stat->mlx, col), "x");
+			i++;
+		}
+
+		x++;
+	}
+}
